# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AssetResponse:
    properties:
      asset_type:
        type: string
      billing_cycle:
        type: string
      cost_amount:
        type: number
      cost_currency:
        type: string
      created_at:
        type: string
      description:
        type: string
      expire_date:
        type: string
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      provider:
        type: string
      remark:
        type: string
      status:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  dto.CreateAssetRequest:
    properties:
      asset_type:
        enum:
        - domain
        - server
        - docker_service
        - ssl_certificate
        - api_key
        - repository
        - other
        type: string
      billing_cycle:
        maxLength: 20
        type: string
      cost_amount:
        type: number
      cost_currency:
        maxLength: 10
        type: string
      description:
        maxLength: 500
        type: string
      expire_date:
        type: string
      identifier:
        maxLength: 200
        type: string
      name:
        maxLength: 100
        type: string
      provider:
        maxLength: 100
        type: string
      remark:
        maxLength: 500
        type: string
      status:
        enum:
        - active
        - inactive
        - expired
        - warning
        type: string
      url:
        maxLength: 500
        type: string
    required:
    - asset_type
    - name
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  dto.UpdateAssetRequest:
    properties:
      asset_type:
        enum:
        - domain
        - server
        - docker_service
        - ssl_certificate
        - api_key
        - repository
        - other
        type: string
      billing_cycle:
        maxLength: 20
        type: string
      cost_amount:
        type: number
      cost_currency:
        maxLength: 10
        type: string
      description:
        maxLength: 500
        type: string
      expire_date:
        type: string
      identifier:
        maxLength: 200
        type: string
      name:
        maxLength: 100
        type: string
      provider:
        maxLength: 100
        type: string
      remark:
        maxLength: 500
        type: string
      status:
        enum:
        - active
        - inactive
        - expired
        - warning
        type: string
      url:
        maxLength: 500
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      username:
        type: string
    type: object
  response.PageResult:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Digital asset and subscription management platform API
  title: StackBill API
  version: "1.0"
paths:
  /assets:
    get:
      consumes:
      - application/json
      description: Get paginated list of digital assets with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by asset type (domain, server, docker_service, ssl_certificate,
          api_key, repository, other)
        in: query
        name: asset_type
        type: string
      - description: Filter by status (active, inactive, expired, warning)
        in: query
        name: status
        type: string
      - description: Filter assets expiring within N days
        in: query
        name: expiring_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/dto.AssetResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List assets
      tags:
      - assets
    post:
      consumes:
      - application/json
      description: Create a new digital asset
      parameters:
      - description: Asset data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssetResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create asset
      tags:
      - assets
  /assets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a digital asset by ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete asset
      tags:
      - assets
    get:
      consumes:
      - application/json
      description: Get a single digital asset by its ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssetResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get asset by ID
      tags:
      - assets
    put:
      consumes:
      - application/json
      description: Update an existing digital asset by ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      - description: Asset data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssetResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update asset
      tags:
      - assets
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username, email, and password
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"